+++ src/Phan/Parse/ParseVisitor.php
@@ -25,6 +25,8 @@ use Phan\Language\FQSEN\FullyQualifiedMethodName;
 use Phan\Language\FQSEN\FullyQualifiedPropertyName;
 use Phan\Language\FutureUnionType;
 use Phan\Language\Type;
+use Phan\Language\Type\ArrayShapeType;
+use Phan\Language\Type\ArrayType;
 use Phan\Language\Type\CallableType;
 use Phan\Language\Type\NullType;
 use Phan\Language\Type\StringType;
@@ -458,33 +460,38 @@ class ParseVisitor extends ScopeVisitor
 
             // Look for any @var declarations
             if ($variable = $comment->getVariableList()[$i] ?? null) {
+                $original_union_type = $union_type;
                 // We try to avoid resolving $future_union_type except when necessary,
                 // to avoid issues such as https://github.com/phan/phan/issues/311 and many more.
                 if ($future_union_type !== null) {
                     try {
-                        $union_type = $future_union_type->get();
+                        $original_union_type = $future_union_type->get();
                         // We successfully resolved the union type. We no longer need $future_union_type
                         $future_union_type = null;
                     } catch (IssueException $e) {
                         // Do nothing
                     }
                     if ($future_union_type === null) {
-                        if ($union_type->isType(NullType::instance(false))) {
+                        if ($original_union_type->isType(ArrayShapeType::empty())) {
+                            $union_type = ArrayType::instance(false)->asUnionType();
+                        } elseif ($original_union_type->isType(NullType::instance(false))) {
                             $union_type = UnionType::empty();
+                        } else {
+                            $union_type = $original_union_type;
                         }
                         // Replace the empty union type with the resolved union type.
                         $property->setUnionType($union_type);
                     }
                 }
 
-                if (!$union_type->isType(NullType::instance(false))
-                    && !$union_type->canCastToUnionType($variable->getUnionType())
+                if (!$original_union_type->isType(NullType::instance(false))
+                    && !$original_union_type->canCastToUnionType($variable->getUnionType())
                     && !$property->hasSuppressIssue(Issue::TypeMismatchProperty)
                 ) {
                     $this->emitIssue(
                         Issue::TypeMismatchProperty,
                         $child_node->lineno ?? 0,
-                        (string)$union_type,
+                        (string)$original_union_type,
                         (string)$property->getFQSEN(),
                         (string)$variable->getUnionType()
                     );